+++ crates/hdk/tests/tests/macros/hdk_to_coordinates_not_enum.rs
@@ -1,6 +1,6 @@
 use hdk::prelude::hdi::prelude::*;
 
-#[hdk_to_coordinates]
+#[hdi_to_coordinates]
 struct Foo;
 
 fn main() {}