+++ crates/test_utils/wasm/wasm_workspace/create_entry/src/integrity.rs
@@ -1,15 +1,15 @@
 use hdi::prelude::*;
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct Post(pub String);
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct Msg(pub String);
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct PrivMsg(pub String);
 
-#[hdk_entry_types]
+#[hdi_entry_types]
 #[unit_enum(EntryTypesUnit)]
 pub enum EntryTypes {
     #[entry_type(required_validations = 5)]
@@ -20,7 +20,7 @@ pub enum EntryTypes {
     PrivMsg(PrivMsg),
 }
 
-#[hdk_link_types]
+#[hdi_link_types]
 pub enum LinkTypes {
     Post,
 }