+++ crates/test_utils/wasm/wasm_workspace/link/src/integrity.rs
@@ -1,15 +1,15 @@
 use hdi::prelude::*;
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct Test;
 
-#[hdk_entry_types]
+#[hdi_entry_types]
 #[unit_enum(EntryTypesUnit)]
 pub enum EntryTypes {
     Test(Test),
 }
 
-#[hdk_link_types]
+#[hdi_link_types]
 pub enum LinkTypes {
     SomeLinks,
     SomeOtherLinks,