+++ crates/test_utils/wasm/wasm_workspace/post_commit_signal/src/integrity.rs
@@ -1,10 +1,10 @@
 use hdk::prelude::*;
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct TestEntry(pub String);
 
 #[derive(Serialize, Deserialize)]
-#[hdk_entry_types]
+#[hdi_entry_types]
 #[unit_enum(UnitEntryTypes)]
 pub enum EntryTypes {
     TestEntry(TestEntry),