+++ crates/test_utils/wasm/wasm_workspace/update_entry/src/integrity.rs
@@ -1,12 +1,12 @@
 use hdi::prelude::*;
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct Post(pub String);
 
-#[hdk_entry_helper]
+#[hdi_entry_helper]
 pub struct Msg(pub String);
 
-#[hdk_entry_types]
+#[hdi_entry_types]
 #[unit_enum(EntryTypesUnit)]
 pub enum EntryTypes {
     #[entry_type(required_validations = 5)]